+++ CHANGELOG.md
@@ -54,6 +54,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - programs: Added `cancel_referral_code_transfer` instruction to cancel a referral code transfer.
 - programs: Added `migrate_referral_code` instruction for `ReferralCode` account migration.
 - programs: Added a new `BorrowingFeesUpdated` CPI event.
+- programs: Added a new `GlvPricing` CPI event.
 - sdk: Added the `gmsol::cli` module.
 - sdk: Added `SwitchboardPullOracleFactory` structure.
 - sdk: Added support for `accept_referral_code` and `cancel_referral_code_transfer` instructions.